+++ cacheback/base.py
@@ -6,7 +6,6 @@ import time
 from django.conf import settings
 from django.core.cache import DEFAULT_CACHE_ALIAS, caches
 from django.db.models import Model as DjangoModel
-from django.utils.itercompat import is_iterable
 
 from .utils import enqueue_task, get_job_class
 
@@ -29,7 +28,7 @@ def to_bytestring(value):
     :returns: a bytestring
     """
     if isinstance(value, DjangoModel):
-        return ('%s:%s' % (value.__class__, hash(value))).encode('utf-8')
+        return (f'{value.__class__}:{hash(value)}').encode('utf-8')
     if isinstance(value, str):
         return value.encode('utf8')
     if isinstance(value, bytes):
@@ -37,7 +36,7 @@ def to_bytestring(value):
     return bytes(str(value), 'utf8')
 
 
-class Job(object):
+class Job:
     """
     A cached read job.
 
@@ -59,7 +58,7 @@ class Job(object):
     #: refresh the cache.
     refresh_timeout = 60
 
-    #: Secifies which cache to use from your `CACHES` setting. It defaults to
+    #: Specifies which cache to use from your `CACHES` setting. It defaults to
     #: `default`.
     cache_alias = None
 
@@ -91,7 +90,7 @@ class Job(object):
 
     @property
     def class_path(self):
-        return '%s.%s' % (self.__module__, self.__class__.__name__)
+        return f'{self.__module__}.{self.__class__.__name__}'
 
     def __init__(self):
         self.cache_alias = self.cache_alias or getattr(
@@ -399,11 +398,11 @@ class Job(object):
             return self.class_path
         try:
             if args and not kwargs:
-                return "%s:%s" % (self.class_path, self.hash(args))
+                return f"{self.class_path}:{self.hash(args)}"
             # The line might break if your passed values are un-hashable.  If
             # it does, you need to override this method and implement your own
             # key algorithm.
-            return "%s:%s:%s:%s" % (
+            return "{}:{}:{}:{}".format(
                 self.class_path,
                 self.hash(args),
                 self.hash([k for k in sorted(kwargs)]),
@@ -422,7 +421,7 @@ class Job(object):
 
         This is for use in a cache key.
         """
-        if is_iterable(value):
+        if isinstance(value, collections.abc.Iterable):
             value = tuple(to_bytestring(v) for v in value)
         return hashlib.md5(b':'.join(value)).hexdigest()
 
@@ -440,7 +439,7 @@ class Job(object):
         :param result: The result to be returned
         :param call: A named tuple with properties 'args' and 'kwargs that
                      holds the call args and kwargs
-        :param cache_status: A status integrer, accessible as class constants
+        :param cache_status: A status integer, accessible as class constants
                              self.MISS, self.HIT, self.STALE
         :param sync_fetch: A boolean indicating whether a synchronous fetch was
                            performed. A value of None indicates that no fetch