+++ docs/conf.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # django-cacheback documentation build configuration file, created by
 # sphinx-quickstart on Mon Jul 30 21:40:46 2012.
@@ -50,8 +49,8 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'django-cacheback'
-copyright = u'2012, David Winterbottom'
+project = 'django-cacheback'
+copyright = '2012, David Winterbottom'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -194,8 +193,8 @@ latex_elements = {
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-  ('index', 'django-cacheback', u'django-cacheback Documentation',
-   u'David Winterbottom', 'manual'),
+  ('index', 'django-cacheback', 'django-cacheback Documentation',
+   'David Winterbottom', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -224,8 +223,8 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index', 'django-cacheback', u'django-cacheback Documentation',
-     [u'David Winterbottom'], 1)
+    ('index', 'django-cacheback', 'django-cacheback Documentation',
+     ['David Winterbottom'], 1)
 ]
 
 # If true, show URL addresses after external links.
@@ -238,8 +237,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index', 'django-cacheback', u'django-cacheback Documentation',
-   u'David Winterbottom', 'django-cacheback', 'One line description of project.',
+  ('index', 'django-cacheback', 'django-cacheback Documentation',
+   'David Winterbottom', 'django-cacheback', 'One line description of project.',
    'Miscellaneous'),
 ]