+++ tests/conftest.py
@@ -10,7 +10,7 @@ def skip_if_no_redis():
         try:
             redis.StrictRedis(
                 settings.RQ_QUEUES['default'].get('HOST', 'localhost'),
-                settings.RQ_QUEUES['default'].get('POST', 6379),
+                settings.RQ_QUEUES['default'].get('PORT', 6379),
             ).ping()
             skip_if_no_redis._redis_available = True
         except redis.ConnectionError:
@@ -36,7 +36,7 @@ def cleared_cache(request):
     cache.clear()
 
 
-@pytest.yield_fixture
+@pytest.fixture()
 def rq_worker(request):
     [queue.empty() for queue in django_rq.get_worker().queues]
 
@@ -47,7 +47,7 @@ def rq_worker(request):
     [queue.empty() for queue in django_rq.get_worker().queues]
 
 
-@pytest.yield_fixture
+@pytest.fixture()
 def rq_burst(request, rq_worker):
     def burst():
         rq_worker.work(burst=True)